(refactor) split partial scalar helpers
